Admin UI/UX separation: introduce dedicated AdminSidebar and sectioned /admin routes to clearly distinguish admin vs chat experiences; default collapsed with persisted state and FR labels/icons
