+++ apps/web/app/admin/settings/page.tsx
@@ -1,0 +1,10 @@
+export default function AdminSettingsPage() {
+  return (
+    <div className="mx-auto w-full max-w-3xl p-6">
+      <h1 className="mb-4 text-2xl font-semibold">Réglages de l’Admin</h1>
+      <div className="rounded-md border p-4">
+        <p className="text-sm text-muted-foreground">Préférences UI de l’espace Admin — à venir.</p>
+      </div>
+    </div>
+  );
+}
