+++ packages/common/components/index.ts
@@ -12,6 +12,7 @@ export * from './history/history-item';
 export * from './icons';
 export * from './intro-dialog';
 export * from './layout/root';
+export * from './layout/admin-sidebar';
 export * from './link-favicon';
 export * from './link-preview';
 export * from './logo';
